fix: js modules and add favicon
