+++ index.html
@@ -10,6 +10,6 @@
                 <tr id='row-3'><td><button ></button></td><td><button ></button></td><td><button ></button></td></tr>
             </tbody>
         </table>
-        <script type="module" src="./main.js" ></script> 
+        <script type="module" src="main.js" ></script> 
     </body>
 </html>
