+++ prettier.config.js
@@ -2,6 +2,6 @@
 module.exports = {
     trailingComma: 'es5',
     tabWidth: 4,
-    semi: false,
+    semi: true,
     singleQuote: true,
 }
